+++ mvp_onboarding.client/src/services/customerService.js
@@ -21,8 +21,13 @@ export const getCustomers = async (pageNumber, pageSize, sortColumn, sortDirecti
 // Get customer 
 export const getCustomer = async (id) => {
   try {
-    const response = await api.get(`/Customer/${id}`);
-    return response.data;
+    if (!(Number.isInteger(id) && id >0)) {
+      throw "Invalid customer id.";
+    }
+    else {
+      const response = await api.get(`/Customer/${id}`);
+      return response.data;
+    }
   } catch (error) {
     console.error(`Error getting customer with ID ${id}:`, error);
     throw error;
@@ -43,8 +48,13 @@ export const createCustomer = async (customerData) => {
 // Update customer
 export const updateCustomer = async (id, customerData) => {
   try {
-    const response = await api.put(`/Customer/${id}`, customerData);
-    return response.data;
+    if (!(Number.isInteger(id) && id >0)) {
+        throw "Invalid customer id.";
+      }
+      else {
+        const response = await api.put(`/Customer/${id}`, customerData);
+        return response.data;
+      }
   } catch (error) {
     console.error(`Error updating customer with ID ${id}:`, error);
     throw error;
@@ -54,8 +64,13 @@ export const updateCustomer = async (id, customerData) => {
 // Delete customer
 export const deleteCustomer = async (id) => {
   try {
-    const response = await api.delete(`/Customer/${id}`);
-    return response.data;
+    if (!(Number.isInteger(id) && id >0)) {
+        throw "Invalid customer id.";
+      }
+      else {
+        const response = await api.delete(`/Customer/${id}`);
+        return response.data;
+      }
   } catch (error) {
     console.error(`Error deleting customer with ID ${id}:`, error);
     throw error;
